A minor improvement

+++ src/impl.h
@@ -13,7 +13,7 @@ struct Impl
             return std::hash<T1>()(pair.first) ^ std::hash<T2>()(pair.second);
         }
     };
-    using bpe_ranks_t = std::unordered_map<std::pair<std::string, std::string>, uint32_t, pair_hash>;
+    using bpe_ranks_t = std::unordered_map<std::pair<std::string, std::string>, uint16_t, pair_hash>;
 
     static size_t count(const std::string& text) noexcept;
     static size_t count(const char* data, size_t len) noexcept;
